+++ src/settings/OrderTemplates/OrderTemplatesEditor/OrderTemplatesEditor.js
@@ -310,6 +310,7 @@ class OrderTemplatesEditor extends Component {
                       order={ORDER}
                       required={false}
                       initialValues={initialValues}
+                      change={change}
                     />
                   </Accordion>