+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Support erm 5.0 interface in orders. Refs UIOR-765.
 * Add vendor invoice number to 'Related invoices' section of PO/POL. Refs UIOR-771.
 * Use mod-orders for piece queries. Refs UIOR-774.
+* Fiscal year rollover amount not shown on POL after rollover. Refs UIOR-768.
 * global CSS styles force landscape printing in other modules. Refs UIOR-775.
 
 ## [2.4.1](https://github.com/folio-org/ui-orders/tree/v2.4.1) (2021-07-28)